Additional updates after 1.0.0 version release.
Keep the version because you never deployed it.

+++ bin/image_dangling_delete.sh
@@ -2,4 +2,6 @@
 
 docker image ls -f dangling=true
 
-docker image rm $(docker image ls -f dangling=true -q)
+if [ `docker image ls -f dangling=true -q | wc -l` != 0 ]; then
+	docker image rm $(docker image ls -f dangling=true -q)
+fi
